+++ src/cruiz/commands/context.py
@@ -24,12 +24,15 @@ from .metarequestconaninvocation import MetaRequestConanInvocation
 
 if typing.TYPE_CHECKING:
     from cruiz.interop.packagenode import PackageNode
-    from cruiz.interop.pod import ConanRemote, ConanHook
+    from cruiz.interop.pod import ConanRemote
     from cruiz.interop.searchrecipesparameters import SearchRecipesParameters
     from cruiz.interop.reciperevisionsparameters import RecipeRevisionsParameters
     from cruiz.interop.packageidparameters import PackageIdParameters
     from cruiz.interop.packagerevisionsparameters import PackageRevisionsParameters
     from cruiz.interop.packagebinaryparameters import PackageBinaryParameters
+
+    from cruizlib.interop.pod import ConanHook
+
     from .conanconf import ConanConfigBoolean
     from .logdetails import LogDetails
