+++ src/cruiz/interop/pod.py
@@ -4,7 +4,6 @@
 
 from __future__ import annotations
 
-import pathlib
 import typing
 from dataclasses import dataclass, field
 
@@ -64,37 +63,6 @@ class ConanRemote:
         url = url_arg[1][1:-1]
         enabled = _strtobool(enabled_arg[1])
         return cls(name, url, enabled)
-
-
-@dataclass(frozen=True)
-class ConanHook:
-    """Plain old data class representing a Conan hook."""
-
-    path: pathlib.Path
-    enabled: bool
-
-    def has_path(self, path: pathlib.Path) -> bool:
-        """Return whether the hook contains the specified path."""
-        return self.path in (path, path.stem)
-
-    @classmethod
-    def from_string(cls, string: str) -> ConanHook:
-        """Convert a string to a ConanHook."""
-        assert string.startswith("ConanHook(")
-        assert string.endswith(")")
-        string = string.replace("ConanHook(", "")
-        string = string.rstrip(")")
-        args = string.split(",")
-        assert len(args) == 2
-        # two args, path=ClassName('/path/to'), enabled=True|False
-        path_arg = args[0].strip().replace("path=", "")
-        path_start = path_arg.find("(") + 1
-        path_end = path_arg.find(")")
-        path_arg = path_arg[path_start:path_end][1:-1]
-        path = pathlib.Path(path_arg)
-        enabled_arg = args[1].strip().replace("enabled=", "")
-        enabled = _strtobool(enabled_arg)
-        return cls(path, enabled)
 
 
 @dataclass(frozen=True)
