+++ src/cruizlib/__init__.py
@@ -1,0 +1,1 @@
+"""Library for non-GUI components of cruiz."""
