+++ src/cruizlib/interop/__init__.py
@@ -1,0 +1,1 @@
+"""Utilities for data interop between Conan and the UI."""
